@@ -1,18 +1,36 @@
 module github.com/bep/s3deploy/v2
 
-go 1.14
+go 1.18
 
 require (
 	github.com/aws/aws-sdk-go-v2 v1.16.14
 	github.com/aws/aws-sdk-go-v2/credentials v1.12.18
 	github.com/aws/aws-sdk-go-v2/service/cloudfront v1.20.3
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.27.9
-	github.com/aws/smithy-go v1.13.2 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/dsnet/golib v0.0.0-20171103203638-1ea166775780
 	github.com/frankban/quicktest v1.14.3
-	golang.org/x/net v0.0.0-20220909164309-bea034e7d591 // indirect
+	github.com/oklog/ulid/v2 v2.1.0
+	github.com/rogpeppe/go-internal v1.9.0
 	golang.org/x/sync v0.0.0-20180314180146-1d60e4601c6f
 	golang.org/x/text v0.3.7
 	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	github.com/aws/aws-sdk-go-v2/aws/protocol/eventstream v1.4.7 // indirect
+	github.com/aws/aws-sdk-go-v2/internal/configsources v1.1.21 // indirect
+	github.com/aws/aws-sdk-go-v2/internal/endpoints/v2 v2.4.15 // indirect
+	github.com/aws/aws-sdk-go-v2/internal/v4a v1.0.12 // indirect
+	github.com/aws/aws-sdk-go-v2/service/internal/accept-encoding v1.9.8 // indirect
+	github.com/aws/aws-sdk-go-v2/service/internal/checksum v1.1.16 // indirect
+	github.com/aws/aws-sdk-go-v2/service/internal/presigned-url v1.9.15 // indirect
+	github.com/aws/aws-sdk-go-v2/service/internal/s3shared v1.13.15 // indirect
+	github.com/aws/smithy-go v1.13.2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/google/go-cmp v0.5.8 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	github.com/kr/pretty v0.3.0 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/pkg/diff v0.0.0-20210226163009-20ebb0f2a09e // indirect
+	golang.org/x/net v0.0.0-20220909164309-bea034e7d591 // indirect
 )
