@@ -1,7 +1,7 @@
 package lib
 
 import (
-	"io/ioutil"
+	"io"
 	"testing"
 
 	qt "github.com/frankban/quicktest"
@@ -17,7 +17,7 @@ func TestNewRemoteStoreNoAclProvided(t *testing.T) {
 		Silent:     true,
 	}
 
-	s, err := newRemoteStore(cfg, newPrinter(ioutil.Discard))
+	s, err := newRemoteStore(cfg, newPrinter(io.Discard))
 	c.Assert(err, qt.IsNil)
 
 	c.Assert("private", qt.Equals, s.acl)
@@ -33,7 +33,7 @@ func TestNewRemoteStoreAclProvided(t *testing.T) {
 		Silent:     true,
 	}
 
-	s, err := newRemoteStore(cfg, newPrinter(ioutil.Discard))
+	s, err := newRemoteStore(cfg, newPrinter(io.Discard))
 	c.Assert(err, qt.IsNil)
 
 	c.Assert("public-read", qt.Equals, s.acl)
@@ -49,7 +49,7 @@ func TestNewRemoteStoreOtherCannedAclProvided(t *testing.T) {
 		Silent:     true,
 	}
 
-	s, err := newRemoteStore(cfg, newPrinter(ioutil.Discard))
+	s, err := newRemoteStore(cfg, newPrinter(io.Discard))
 	c.Assert(err, qt.IsNil)
 
 	c.Assert("bucket-owner-full-control", qt.Equals, s.acl)
@@ -66,7 +66,7 @@ func TestNewRemoteStoreDeprecatedPublicReadACLFlaglProvided(t *testing.T) {
 		Silent:        true,
 	}
 
-	s, err := newRemoteStore(cfg, newPrinter(ioutil.Discard))
+	s, err := newRemoteStore(cfg, newPrinter(io.Discard))
 	c.Assert(err, qt.IsNil)
 
 	c.Assert("public-read", qt.Equals, s.acl)
